requirements.txt
numpy
scipy
scikit-learn
pandas
matplotlib
seaborn
Pillow
Django